refactor: consolidate AgentCapabilities to single canonical type (Phase 02)
Fix a type-shadowing bug in src/renderer/global.d.ts where
AgentCapabilities was declared twice in the same file (once at line 90
with supportsThinkingDisplay, again at line 133 without), causing the
second declaration to silently override the first.

Consolidate all 6 duplicate AgentCapabilities definitions in the
codebase to a single canonical interface in src/shared/types.ts along
with a canonical DEFAULT_CAPABILITIES constant. The canonical shape is
a strict superset of every prior definition, including:

- All 24 boolean capability flags
- Optional imageResumeMode field (formerly only in main/agents/capabilities)
- Full JSDoc on every field

Files updated to import from the canonical source:
- src/main/agents/capabilities.ts: re-exports from shared/types
- src/main/preload/agents.ts: drops local definition, imports from shared
- src/renderer/global.d.ts: both definitions replaced with import() alias
- src/renderer/types/index.ts: re-exports from shared/types
- src/renderer/hooks/agent/useAgentCapabilities.ts: imports from shared

Also adds a missing !! coercion in useAgentCapabilities.hasCapability
since the canonical type includes imageResumeMode (string | undefined),
which breaks the strict boolean return type when indexed by key.

Verified: lint clean (all three tsconfigs), capability tests pass
(228/228 across capabilities, agent-completeness, detector,
useAgentCapabilities, useAvailableAgents, ipc/handlers/agents).
